Revert "Support PSR-4 autoloading with Composer"
This reverts commit b1ac99e3fdf3f44bc299f3079f2b17a615a42892.
