--- composer.json
@@ -1,12 +1,0 @@
-{
-    "name": "fatzebra/fatzebra",
-    "description": "Fat Zebra Australian online payment gateway library",
-    "require": {
-        "php": ">=5.3.0"
-    },
-    "autoload": {
-        "psr-4": {
-            "FatZebra\\": "src"
-        }
-    }
-}
