--- src/TimeoutException.php
@@ -1,21 +1,0 @@
-<?php
-	/**
-	* Fat Zebra PHP Gateway Library
-	* Version 1.1.5
-	*
-	* The original source for this library, including its tests can be found at
-	* https://github.com/fatzebra/PHP-Library
-	*
-	* Please visit http://docs.fatzebra.com.au for details on the Fat Zebra API
-	* or https://www.fatzebra.com.au/help for support.
-	*
-	* Patches, pull requests, issues, comments and suggestions always welcome.
-	*
-	* @package FatZebra
-	*/
-	namespace FatZebra;
-
-	/**
-	* The Fat Zebra Gateway class for interfacing with Fat Zebra
-	*/
-	class TimeoutException extends \Exception {}
